requirements.txt
pagarme-python==4.0.2
pydantic
python-dotenv